requirements.txt
attrs==20.2.0
bcrypt==3.2.0
cffi==1.14.2
cryptography==3.1
falcon==2.0.0
gunicorn==20.0.4
iniconfig==1.0.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
py-eureka-client==0.8.2
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
six==1.15.0
toml==0.10.1